📝 move lexakai folders into internal package to clarify that documentation is not exported  - remove lexakai package exports
