+++ kivakit-filesystems/s3fs/src/main/java/com/telenav/kivakit/filesystems/s3fs/S3File.java
@@ -23,7 +23,7 @@ import com.telenav.kivakit.core.time.Time;
 import com.telenav.kivakit.core.value.count.Bytes;
 import com.telenav.kivakit.filesystem.FilePath;
 import com.telenav.kivakit.filesystem.spi.FileService;
-import com.telenav.kivakit.filesystems.s3fs.lexakai.DiagramS3;
+import com.telenav.kivakit.filesystems.s3fs.internal.lexakai.DiagramS3;
 import com.telenav.lexakai.annotations.LexakaiJavadoc;
 import com.telenav.lexakai.annotations.UmlClassDiagram;
 import software.amazon.awssdk.services.s3.model.GetObjectRequest;
