+++ kivakit-filesystems/s3fs/src/main/java/com/telenav/kivakit/filesystems/s3fs/S3FileSystemService.java
@@ -24,7 +24,7 @@ import com.telenav.kivakit.filesystem.spi.DiskService;
 import com.telenav.kivakit.filesystem.spi.FileService;
 import com.telenav.kivakit.filesystem.spi.FileSystemService;
 import com.telenav.kivakit.filesystem.spi.FolderService;
-import com.telenav.kivakit.filesystems.s3fs.lexakai.DiagramS3;
+import com.telenav.kivakit.filesystems.s3fs.internal.lexakai.DiagramS3;
 import com.telenav.kivakit.resource.Resource;
 import com.telenav.kivakit.filesystem.FilePath;
 import com.telenav.lexakai.annotations.LexakaiJavadoc;
