+++ kivakit-filesystems/s3fs/src/main/java/com/telenav/kivakit/filesystems/s3fs/S3Folder.java
@@ -27,7 +27,7 @@ import com.telenav.kivakit.filesystem.FilePath;
 import com.telenav.kivakit.filesystem.Folder;
 import com.telenav.kivakit.filesystem.spi.FileService;
 import com.telenav.kivakit.filesystem.spi.FolderService;
-import com.telenav.kivakit.filesystems.s3fs.lexakai.DiagramS3;
+import com.telenav.kivakit.filesystems.s3fs.internal.lexakai.DiagramS3;
 import com.telenav.kivakit.interfaces.comparison.Matcher;
 import com.telenav.kivakit.resource.FileName;
 import com.telenav.lexakai.annotations.LexakaiJavadoc;
