+++ kivakit-filesystems/s3fs/src/main/java/module-info.java
@@ -15,5 +15,4 @@ open module kivakit.filesystems.s3fs
     
     // Module exports
     exports com.telenav.kivakit.filesystems.s3fs;
-    exports com.telenav.kivakit.filesystems.s3fs.lexakai;
 }
