+++ kivakit-logs/email/src/main/java/module-info.java
@@ -10,5 +10,4 @@ open module kivakit.logs.email
 
     // Module exports
     exports com.telenav.kivakit.logs.email;
-    exports com.telenav.kivakit.logs.email.lexakai;
 }
