+++ kivakit-microservice/src/main/java/com/telenav/kivakit/microservice/internal/lexakai/DiagramJetty.java
@@ -16,13 +16,13 @@
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-package com.telenav.kivakit.microservice.lexakai;
+package com.telenav.kivakit.microservice.internal.lexakai;
 
 import com.telenav.lexakai.annotations.diagrams.UmlDiagramIdentifier;
 
 /**
  * @author jonathanl (shibo)
  */
-public class DiagramMicroservlet implements UmlDiagramIdentifier
+public class DiagramJetty implements UmlDiagramIdentifier
 {
 }
