+++ kivakit-microservice/src/main/java/com/telenav/kivakit/microservice/internal/lexakai/DiagramMicroservlet.java
@@ -16,13 +16,13 @@
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 
-package com.telenav.kivakit.logs.file.lexakai;
+package com.telenav.kivakit.microservice.internal.lexakai;
 
 import com.telenav.lexakai.annotations.diagrams.UmlDiagramIdentifier;
 
 /**
  * @author jonathanl (shibo)
  */
-public class DiagramLogsFile implements UmlDiagramIdentifier
+public class DiagramMicroservlet implements UmlDiagramIdentifier
 {
 }
