+++ kivakit-microservice/src/main/java/com/telenav/kivakit/microservice/microservlet/BaseMicroservletRequest.java
@@ -3,7 +3,7 @@ package com.telenav.kivakit.microservice.microservlet;
 import com.telenav.kivakit.component.BaseComponent;
 import com.telenav.kivakit.core.time.Duration;
 import com.telenav.kivakit.core.time.Frequency;
-import com.telenav.kivakit.microservice.lexakai.DiagramMicroservlet;
+import com.telenav.kivakit.microservice.internal.lexakai.DiagramMicroservlet;
 import com.telenav.kivakit.validation.ValidationType;
 import com.telenav.kivakit.validation.Validator;
 import com.telenav.lexakai.annotations.UmlClassDiagram;
