+++ kivakit-microservice/src/main/java/com/telenav/kivakit/microservice/microservlet/MicroservletResponse.java
@@ -2,7 +2,7 @@ package com.telenav.kivakit.microservice.microservlet;
 
 import com.telenav.kivakit.component.Component;
 import com.telenav.kivakit.microservice.internal.protocols.rest.cycle.HttpProblemReportingTrait;
-import com.telenav.kivakit.microservice.lexakai.DiagramMicroservlet;
+import com.telenav.kivakit.microservice.internal.lexakai.DiagramMicroservlet;
 import com.telenav.kivakit.validation.Validatable;
 import com.telenav.lexakai.annotations.UmlClassDiagram;
